+++ shell/components/RelatedWorkloadsTable.vue
@@ -40,5 +40,8 @@ export default {
 </script>
 
 <template>
+<div>
+  RESOURCE TABLE
   <ResourceTable :schema="schema" :rows="relatedWorkloadRows" :headers="relatedWorkloadHeaders" />
+</div>
 </template>
